% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qecol.R
\name{q_regres}
\alias{q_regres}
\title{Quickly run a regression with one command.}
\usage{
q_regres(x, y, xlab, ylab)
}
\arguments{
\item{x}{Numerical vector.}

\item{y}{A Numerical vector.}

\item{xlab}{Label for x axis.}

\item{ylab}{Label for y axis.}
}
\value{
Runs a quick regression and prints the results to a markdown report.
}
\description{
Quickly run a regression with one command.
}
\examples{
data(mussels)
q_regres(mussels$Lshell,mussels$BTVolume,"Shell length","Body volume")
}

